[Feat] - 결과/스탯 및 아이템 페이지 컴포넌트 분리
